Fixed stack reference map preparation for a thread stopped at a safepoint.
The reference map for the stack of a thread stopped at a safepoint did not
take into the account the frames created below the frame in which the
safepoint trap occurred. These extra frames correspond to the code path
that ends with a call in native code to block the thread after preparing
the reference map for the frames above and including the trap frame.