Fixed bug in stack reference map scanning.
The traversal of the reference map was at byte-level granularity.
That is, bits in the reference map outside of the bit range were
being scanned due to alignment issues. The scanning is now
exact with respect to the requested bit range.

Code specific to a particular GC scheme (Beltwa)
was removed from the shared GC code.