Fix implementation of new Trap mechanism for SPARC.
The trap state for SPARC comprises only the floating point registers and the
general purpose registers that:
- are not in the trapped frame register windows saving area
- are not system-reserved global registers (namely, g0, g6, g7)

That is, the trap state only includes g2 to g5 plus all the %o registers -- which
happen to be %i in the trapStub frame...
Further, we do not push anything on the stack, especially NOT the safepoint register.
Instead, we change its value directly in the ucontext in trap.c.
This could be done for AMD64 as well, it would simplify further the trapping / safepoint mechanism.