Added -XX:ExcessiveGCFrequency=<n> VM option that runs a garbage collection every <n> milliseconds. A value of 0 (the default) disables this mechanism.
Normalized all the naming of lock word variables and classes to not camel-case the 'w'. That is, everything is now named 'lockword' or 'Lockword'.

Changed some javadoc in the non-C1X projects to replace '<code>...</code>' with '{@code ...}'.

Simplified the methods in Log.java for printing methods/fields/threads.

Fixed bug in StandardJavaMonitor where the binding protection of a monitor was being set to UNPROTECTED on a monitorexit operation thast released the lock (i.e. recursion count == 0). The binding protection should not be transitioned to UNPROTECTED if there are threads waiting on the monitor otherwise the invariant that all monitors that have a non-null owner must be PROTECTED is violated when one of the waiting threads wakes up.