Additions to get NetBeans running: - Skeleton implementation of JMM native interface to get NetBeans execution working. - Implemented more of JVM interface (i.e. jvm.c).
Fixed bugs preventing a JavaMethodInspector from working when inspecting a method for which there is only bytecode, no target code.

Removed offset to 'info' field in Inspectable.java from boot image header. This field is now named 'flags' and has been generalized to be a mask of flags that modify the VMs execution according to certain properties of the Inspector. For example, the garbage collector informs the Inspector of each object relocation so that the Inspector can relocate any watchpoint(s) associated with the object. Using the new flags, the garbage collector only performs this notification if the new ACTIVE_RELOCATABLE_WATCHPOINTS flag is set.

Added ability to disable growing the heap (-XX:HeapGrowPolicy=None).