Completed implementation of pinning a variable to the stack and exposing its address. The register allocator extends the lifetime range of the interval corresponding to the pinned variable to the end of the method so that it is never reused. It also never spilts such an interval which ensures it will always be on the stack.
Revived C1 optimization that elides a block at bytecode parsing time when inlining a method that only has a single block delimited by a return (e.g. setter and getter methods). This optimization was previously crippled in C1X.

Implemented constant folding for non-static fields where the receiver object is known at compile time.

JNI stub tracing in boot image code is now only enabled for DEBUG builds of the VM.

Created CiUtil.java and moved some methods from Util.java to it for use by other Ci* classes.

Broadened @NEVER_INLINE so that it can be applied to a class. This is equivalent to applying it to all the methods in the class *except<* for those explicitly annotated with @INLINE.